feat: add new hook useDynamicFeeOptions
